requirements.txt
numpy
natsort
matplotlib
PyQt5
pyqtgraph
PyOpenGL
python-lzf
scipy
pillow
setuptools
pyinstaller
